demoui/src/main/java/demoui/domain/User.java
package demoui.domain;

import java.io.Serializable;

public class User implements Serializable {

    public String getFamilyName() {
        return familyName;
    }

    public void setFamilyName(String familyName) {
        this.familyName = familyName;
    }

    public String getGivenName() {
        return givenName;
    }

    public void setGivenName(String givenName) {
        this.givenName = givenName;
    }

    public String getUserName() {
        return userName;
    }

    public void setUserName(String userName) {
        this.userName = userName;
    }

    private String familyName;
    private String givenName;
    private String userName;


}